Major item update
Items now use "stat effects" to grant stats. Every effect has a name, stat to give and amount of stat.  Effects with name different from "None" are counted as a unique effect and will be applied once only.

The amount of stats that an item can grant is infinite and may grant the same stat more than once, including with Unique effects. The rest of the item should be acting as usual.

Status effects may now give stats the same way as items do.

Shop was changed to suit the new item system, but works pretty much the same.

ShinbiTest now has Mana instead of Fury for testing purposes.

Knoife with Bipod now grants 10 mana regen for testing purposes.

Helicopter Blades now grant 100 max mana for testing purposes.

Removed GameplayAbilities plugin as the systems in this project are doing the same work, but for less hassle.