Major shop UI update
Items are no longer in lines, they're now in a grid.

Clicking a grid item will show the recipe on the right and you may click an item inside the recipe to show it separately.

[Maybe an old change] You may now sell an item using the "Sell" button.

Side updates:

New function in the function library to get the item array of a unit.

TestMap was updated a tiny bit to update the shop.