Item update
Items now have additional stats:
- Magic and armor penetration (including %)
- Crit chance

Added a bunch of items from the sheet, each with a respective texture:
- Dagger
- Sealed Tome
- Gunwand
- Chainmail
- Leather Tunic
- Leech (difference from sheet - spellvamp is unique)
- Musket
- Elemental Flower
- Shiv
- Silk Cloak
- Essence Pearl
- Earth Trinket
- Soulstone
- Scimitar (difference from sheet - crit chance is unique)

Moved "inactive" assets from BPs into a separate folder to remove clutter (assets like structs, enums).

Shop contents now have a scale of 1 (up from 0.8) to avoid unused space.

Abilities are now in an array instead of a map to aid in replication, but is still in progress.