Eminusk prototype
Reworked projectiles to use an event on top of a function to let other actors (units/status effects) listen for when it hits/reaches max distance.

Fixed sphere targeting heading towards center of the map if not aimed directly at ground.

Added "Add/Take Resource" to MOBAHero to handle resource management in a more consistent and clean way.

Added prototypes for Eminusk.

Added "no mana (or resource)" indicator to abilities.